♻️Refactor : UICollectionViewCell의 extension을 OpenMarketCell로 이동
